Added a proper startConnection function
This will start the actual connection and put the initial data in
the buffer. So the basic stuff like password, nickname and the username.
Also added a simple parser for the numbered headers, which I do not yet
properly handle, but this is a start of that. And as I mentioned password
earlier you now have the option to set a password in the config file under
the key password. Which is of course the connection password.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>